+++ src/components/MyCreditRechargeButton.module.css
@@ -1,6 +1,6 @@
 .MycreditButton {
   position: relative;
-  width: 327px !important;
+  width: 100% !important;
   height: 87px !important;
   border: 1px solid #f1eef9cc !important;
   border-radius: 8px !important;
