+++ src/pages/ListPage/ListPage.module.css
@@ -1,0 +1,30 @@
+.listWrapper{
+    width: 100%;
+}
+
+.rechargeButtonwrapper{
+    width: 100%;
+    padding: 0 24px;
+}
+
+.creditRechargeButton {
+    margin-top: 16px
+}
+
+.carouselSectionWrapper{
+    margin-top: 40px;
+    margin-left: 24px;
+    height: 303px;
+}
+
+.chartSectionWrapper{
+    margin: 40px 24px 59px;
+}
+
+.chartSection{
+    display: block;
+    width: 100%;
+    height: 553px;
+    background-color: var(--mantine-color-white-0);
+}
+
